Repo cleanup
Some repo maintenance:
- moving all config files to a config directory
- moving scripts and integration test to the tools directory
- removed go example using opentelemetry lightstep exporter
- renamed go/otlp to go/opentelemetry
- renamed python/otlp to python/opentelemetry
- update README
