+++ .github/workflows/integration.yml
@@ -22,10 +22,10 @@ jobs:
         ORG_NAME: ${{ secrets.ORG_NAME }}
         PROJECT_NAME: ${{ secrets.PROJECT_NAME }}
         API_KEY: ${{ secrets.API_KEY }}
-      run: ./scripts/update-token.sh
+      run: ./tools/update-token.sh
 
     - name: Validate config is up-to-date
-      run: ./scripts/precommit.sh
+      run: ./tools/precommit.sh
 
     - name: Build containers
       run: docker-compose build
