+++ .gitignore
@@ -12,7 +12,7 @@ target
 dependency-reduced-pom.xml
 *.jar
 .swp
-collector/collector-config.yaml
+config/collector-config.yaml
 
 dotnet/server/aspnetapp/bin
 dotnet/server/aspnetapp/obj
