--- collector/README.md
@@ -1,3 +1,0 @@
-# Collector configuration
-
-This folder will contain the collector configuration used when launching the `otel-collector` via docker-compose. See the [README.md](../README.md) for more details.
