+++ config/example-collector-config.yaml
