+++ config/integration.yml
@@ -1,10 +1,10 @@
 clients:
 - go-opentracing-client
-- go-otlp-client
+- go-opentelemetry-client
 - go-launcher-client
 - py-lstrace-client
 - py-collector-client
-- py-otlp-client
+- py-opentelemetry-client
 - py-launcher-client
 - py-opentracing-client
 - js-lstrace-client
@@ -15,11 +15,11 @@ clients:
 - java-launcher-client
 endpoints:
 - http://go-opentracing-server:8081/ping
-- http://go-otlp-server:8081/ping
+- http://go-opentelemetry-server:8081/ping
 - http://go-launcher-server:8081/ping
 - http://py-lstrace-server:5000/ping
 - http://py-collector-server:5000/ping
-- http://py-otlp-server:5000/ping
+- http://py-opentelemetry-server:5000/ping
 - http://py-launcher-server:5000/ping
 - http://py-opentracing-server:5000/ping
 - http://js-lstrace-server:8080/ping
@@ -32,11 +32,11 @@ endpoints:
 - http://dotnet-otlp-server
 services:
 - go-opentracing-server
-- go-otlp-server
+- go-opentelemetry-server
 - go-launcher-server
 - py-lstrace-server
 - py-collector-server
-- py-otlp-server
+- py-opentelemetry-server
 - py-launcher-server
 - py-opentracing-server
 - js-lstrace-server
