+++ docker-compose.yml
@@ -28,9 +28,9 @@ services:
             - LS_SERVICE_VERSION=1.2.7
             - DESTINATION_URL=http://go-opentracing-server:8081/ping
         stop_grace_period: 1s
-    go-otlp-server:
-        container_name: go-otlp-server
-        build: ./go/otlp/server
+    go-opentelemetry-server:
+        container_name: go-opentelemetry-server
+        build: ./go/opentelemetry/server
         networks:
             - demo
         env_file:
@@ -40,25 +40,25 @@ services:
         environment:
             - LS_SATELLITE_URL=otel-collector:55680
             - LS_INSECURE=1
-            - LS_SERVICE_NAME=go-otlp-server
+            - LS_SERVICE_NAME=go-opentelemetry-server
             - LS_SERVICE_VERSION=9.9.1
         stop_grace_period: 1s
-    go-otlp-client:
-        container_name: go-otlp-client
-        build: ./go/otlp/client
+    go-opentelemetry-client:
+        container_name: go-opentelemetry-client
+        build: ./go/opentelemetry/client
         networks:
             - demo
         depends_on:
-            - go-otlp-server
+            - go-opentelemetry-server
             - otel-collector
         env_file:
             - .env
         environment:
             - LS_SATELLITE_URL=otel-collector:55680
             - LS_INSECURE=1
-            - LS_SERVICE_NAME=go-otlp-client
+            - LS_SERVICE_NAME=go-opentelemetry-client
             - LS_SERVICE_VERSION=9.9.2
-            - DESTINATION_URL=http://go-otlp-server:8081/ping
+            - DESTINATION_URL=http://go-opentelemetry-server:8081/ping
         stop_grace_period: 1s
     go-launcher-server:
         container_name: go-launcher-server
@@ -120,7 +120,7 @@ services:
     py-collector-client:
         container_name: py-collector-client
         build:
-            context: ./python/otlp
+            context: ./python/opentelemetry
             dockerfile: Dockerfile.client
         networks:
             - demo
@@ -139,7 +139,7 @@ services:
     py-collector-server:
         container_name: py-collector-server
         build:
-            context: ./python/otlp
+            context: ./python/opentelemetry
             dockerfile: Dockerfile.server
         networks:
             - demo
@@ -153,27 +153,27 @@ services:
             - LS_SERVICE_NAME=py-collector-server
             - LS_SERVICE_VERSION=0.2.8
         stop_grace_period: 1s
-    py-otlp-client:
-        container_name: py-otlp-client
+    py-opentelemetry-client:
+        container_name: py-opentelemetry-client
         build:
-            context: ./python/otlp
+            context: ./python/opentelemetry
             dockerfile: Dockerfile.client
         networks:
             - demo
         depends_on:
-            - py-otlp-server
+            - py-opentelemetry-server
         env_file:
             - .env
         environment:
             - COLLECTOR_ENDPOINT=ingest.lightstep.com:443
-            - LS_SERVICE_NAME=py-otlp-client
+            - LS_SERVICE_NAME=py-opentelemetry-client
             - LS_SERVICE_VERSION=10.10.10
-            - DESTINATION_URL=http://py-otlp-server:5000/ping
+            - DESTINATION_URL=http://py-opentelemetry-server:5000/ping
         stop_grace_period: 1s
-    py-otlp-server:
-        container_name: py-otlp-server
+    py-opentelemetry-server:
+        container_name: py-opentelemetry-server
         build:
-            context: ./python/otlp
+            context: ./python/opentelemetry
             dockerfile: Dockerfile.server
         networks:
             - demo
@@ -181,7 +181,7 @@ services:
             - .env
         environment:
             - COLLECTOR_ENDPOINT=ingest.lightstep.com:443
-            - LS_SERVICE_NAME=py-otlp-server
+            - LS_SERVICE_NAME=py-opentelemetry-server
             - LS_SERVICE_VERSION=10.10.9
         stop_grace_period: 1s
     py-launcher-client:
@@ -227,7 +227,7 @@ services:
             - LS_SERVICE_NAME=demo-client-otlp
             - LS_SERVICE_VERSION=0.1.1
         volumes:
-            - ./integration:/config/:ro
+            - ./config:/config/:ro
         stop_grace_period: 1s
     demo-client-lstrace:
         container_name: demo-client-lstrace
@@ -243,7 +243,7 @@ services:
             - LS_SERVICE_NAME=demo-client-lstrace
             - LS_SERVICE_VERSION=0.2.2
         volumes:
-            - ./integration:/config/:ro
+            - ./config:/config/:ro
         stop_grace_period: 1s
     py-opentracing-client:
         container_name: py-opentracing-client
@@ -476,16 +476,15 @@ services:
     integration-test:
         container_name: integration-test
         build:
-            context: ./integration
+            context: ./tools/integration
         networks:
             - demo
         depends_on:
             - go-opentracing-server
-            - go-otlp-server
+            - go-opentelemetry-server
             - go-launcher-server
-            - py-lstrace-server
             - py-collector-server
-            - py-otlp-server
+            - py-opentelemetry-server
             - py-launcher-server
             - py-opentracing-server
             - js-lstrace-server
@@ -502,7 +501,7 @@ services:
             - LS_SERVICE_NAME=integration-test
             - OTEL_RESOURCE_ATTRIBUTES=service.version=2020.10.02
         volumes:
-            - ./integration:/config/:ro
+            - ./config:/config/:ro
         stop_grace_period: 1s
     ################################
     ####### Config generator #######
@@ -510,7 +509,7 @@ services:
     config-generator:
         container_name: config-generator
         build:
-            context: ./integration
+            context: ./tools/integration
         env_file:
             - .env
         environment:
@@ -518,7 +517,7 @@ services:
             - OUTPUT_PATH=/config/integration.yml
         volumes:
             - ./docker-compose.yml:/docker-compose.yml/:ro
-            - ./integration:/config/:rw
+            - ./config:/config/:rw
         command: python /app/generate_config.py
         stop_grace_period: 1s
     ##############################
@@ -574,7 +573,7 @@ services:
         #     - jaeger
         #     - satellite
         volumes:
-            - ./collector/collector-config.yaml:/conf/collector-config.yaml
+            - ./config/collector-config.yaml:/conf/collector-config.yaml
         ports:
             - "0.0.0.0:8889:8889"   # Prometheus exporter metrics
             - "0.0.0.0:55680:55680" # OTLP receiver
