--- go/go.mod
@@ -1,21 +1,0 @@
-module github.com/lightstep/example
-
-go 1.13
-
-require (
-	github.com/bombsimon/wsl v1.2.8 // indirect
-	github.com/fatih/color v1.9.0 // indirect
-	github.com/go-critic/go-critic v0.4.1 // indirect
-	github.com/golangci/golangci-lint v1.22.2 // indirect
-	github.com/lightstep/opentelemetry-exporter-go v0.1.3 // indirect
-	github.com/pelletier/go-toml v1.6.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/securego/gosec v0.0.0-20200106085552-9cb83e10afad // indirect
-	github.com/spf13/cast v1.3.1 // indirect
-	go.opentelemetry.io/otel v0.3.0
-	golang.org/x/sys v0.0.0-20200113162924-86b910548bc1 // indirect
-	golang.org/x/tools v0.0.0-20200115204027-20e5918576e8 // indirect
-	google.golang.org/grpc v1.27.1
-	gopkg.in/ini.v1 v1.51.1 // indirect
-	mvdan.cc/unparam v0.0.0-20191111180625-960b1ec0f2c2 // indirect
-)
