+++ go/opentelemetry/client/go.mod
@@ -3,10 +3,13 @@ module github.com/lightstep/ls-examples/go/opentelemetry/client
 go 1.13
 
 require (
-	github.com/gogo/protobuf v1.3.1 // indirect
 	github.com/kr/pretty v0.2.0 // indirect
-	github.com/lightstep/opentelemetry-exporter-go v0.6.2
-	go.opentelemetry.io/otel v0.6.1-0.20200610051553-e53841a4b4fd
-	google.golang.org/grpc v1.29.1 // indirect
-	gopkg.in/yaml.v2 v2.2.8 // indirect
+	go.opentelemetry.io/contrib/instrumentation/net/http/httptrace/otelhttptrace v0.12.0
+	go.opentelemetry.io/contrib/propagators v0.12.0
+	go.opentelemetry.io/otel v0.12.0
+	go.opentelemetry.io/otel/exporters/otlp v0.12.0
+	go.opentelemetry.io/otel/sdk v0.12.0
+	golang.org/x/sys v0.0.0-20200615200032-f1bc736245b1 // indirect
+	golang.org/x/text v0.3.2 // indirect
+	google.golang.org/grpc v1.32.0
 )
