+++ go/opentelemetry/server/server.go
@@ -14,25 +14,28 @@ import (
 	"log"
 	"math/rand"
 	"net/http"
-	"net/url"
 	"os"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/gorilla/mux"
-
-	ls "github.com/lightstep/opentelemetry-exporter-go/lightstep"
-	muxtrace "go.opentelemetry.io/contrib/instrumentation/gorilla/mux"
+	muxtrace "go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux"
+	"go.opentelemetry.io/contrib/propagators/b3"
 	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/api/propagation"
+	"go.opentelemetry.io/otel/exporters/otlp"
+	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/sdk/resource"
 	"go.opentelemetry.io/otel/sdk/trace"
+	"google.golang.org/grpc/credentials"
 )
 
 var (
-	lsToken        = os.Getenv("LS_ACCESS_TOKEN")
-	lsMetricsURL   = os.Getenv("LS_METRICS_URL")
 	componentName  = os.Getenv("LS_SERVICE_NAME")
 	serviceVersion = os.Getenv("LS_SERVICE_VERSION")
+	lsToken        = os.Getenv("LS_ACCESS_TOKEN")
+	collectorURL   = os.Getenv("LS_SATELLITE_URL")
+	insecure       = os.Getenv("LS_INSECURE")
 )
 
 const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
@@ -62,59 +65,72 @@ func randString(n int) string {
 	return sb.String()
 }
 
-func initLightstepTracer() {
-	u, err := url.Parse(lsMetricsURL)
+func initExporter(url string, token string) *otlp.Exporter {
+	headers := map[string]string{
+		"lightstep-access-token": token,
+	}
 
-	host := "ingest.lightstep.com"
-	port := 443
-	plaintext := false
+	secureOption := otlp.WithTLSCredentials(credentials.NewClientTLSFromCert(nil, ""))
+	if len(insecure) > 0 {
+		secureOption = otlp.WithInsecure()
+	}
 
-	if err == nil {
-		host = u.Hostname()
-		port, _ = strconv.Atoi(u.Port())
-		if u.Scheme == "http" {
-			plaintext = true
-		}
+	exporter, err := otlp.NewExporter(
+		secureOption,
+		otlp.WithAddress(url),
+		otlp.WithHeaders(headers),
+	)
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	return exporter
+}
+
+func initTracer() {
+	b3 := b3.B3{}
+	// Register the B3 propagator globally.
+	global.SetPropagators(propagation.New(
+		propagation.WithExtractors(b3),
+		propagation.WithInjectors(b3),
+	))
+
+	if len(collectorURL) == 0 {
+		collectorURL = "localhost:55680"
 	}
 
 	if len(componentName) == 0 {
 		componentName = "test-go-server"
 	}
+
 	if len(serviceVersion) == 0 {
 		serviceVersion = "0.0.0"
 	}
 
-	exporter, err := ls.NewExporter(
-		ls.WithAccessToken(lsToken),
-		ls.WithHost(host),
-		ls.WithPort(port),
-		ls.WithPlainText(plaintext),
-		ls.WithServiceName(componentName),
-		ls.WithServiceVersion(serviceVersion),
+	exporter := initExporter(collectorURL, lsToken)
+
+	resources := resource.New(
+		label.String("service.name", componentName),
+		label.String("service.version", serviceVersion),
+		label.String("library.language", "go"),
+		label.String("library.version", "1.2.3"),
 	)
-	if err != nil {
-		log.Fatal(err)
-	}
-	tp, err := trace.NewProvider(trace.WithConfig(trace.Config{DefaultSampler: trace.AlwaysSample()}),
-		trace.WithSyncer(exporter))
-	if err != nil {
-		log.Fatal(err)
-	}
-	global.SetTraceProvider(tp)
+	tp := trace.NewTracerProvider(
+		trace.WithConfig(trace.Config{DefaultSampler: trace.AlwaysSample()}),
+		trace.WithSyncer(exporter),
+		trace.WithResource(resources),
+	)
+	global.SetTracerProvider(tp)
 }
 
 func main() {
-	initLightstepTracer()
+	initTracer()
 	fmt.Printf("Starting server on http://localhost:8081\n")
 	r := mux.NewRouter()
+	// re-enable once the new version of otel-go and otel-go-contrib is released
 	r.Use(muxtrace.Middleware(componentName))
-	r.HandleFunc("/content/{length:[0-9]+}", func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		length, err := strconv.Atoi(vars["length"])
-		if err != nil {
-			length = 10
-		}
-
+	r.HandleFunc("/ping", func(w http.ResponseWriter, r *http.Request) {
+		length := rand.Intn(1024)
 		log.Printf("%s %s %s", r.Method, r.URL.Path, r.Proto)
 		fmt.Fprintf(w, randString(length))
 	})
