--- go/otlp/server/Dockerfile
@@ -1,6 +1,0 @@
-FROM golang:latest
-WORKDIR /app
-COPY go.mod go.sum ./
-RUN go mod download
-COPY *.go .
-CMD ["go", "run", "server.go"]
