+++ python/opentelemetry/Dockerfile.client
