+++ python/opentelemetry/common.py
