+++ tools/integration/Dockerfile
