+++ tools/precommit.sh
@@ -4,7 +4,7 @@ docker-compose up config-generator
 
 if ! git diff --quiet; then
     echo;
-    echo 'Working tree is not clean, did you forget to run "./scripts/precommit"?';
+    echo 'Working tree is not clean, did you forget to run "./tools/precommit.sh"?';
     echo;
     git status;
     exit 1;
